+++ src/levels/level.rs
@@ -50,12 +50,12 @@ pub struct LevelConfig {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct EscalatorConfig {
-    pub x: f32,
-    pub y: f32,
-    pub width: f32,
-    pub height: f32,
+    pub x: i32,
+    pub y: i32,
+    pub width: i32,
+    pub height: i32,
     pub num_steps: i32,
-    pub speed: f32,
+    pub speed: i32,
     pub direction: Direction,
     pub color_flag: ColorFlag,
 }
